+++ src/main/java/com/example/gestion-bibliotheque/service/CategorieService.java
@@ -1,8 +1,9 @@
 package com.example.gestion_bibliotheque.service;
 
-import com.example.gestion_bibliotheque.entity.Auteur;
 import com.example.gestion_bibliotheque.entity.Categorie;
+import com.example.gestion_bibliotheque.exception.ResourceNotFoundException;
 import com.example.gestion_bibliotheque.repository.CategorieRepository;
+import com.example.gestion_bibliotheque.utils.FunctionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,23 +24,17 @@ public class CategorieService {
     }
 
     public Categorie findById(int id) {
+        FunctionUtils.validateId(id);
         return categorieRepository.findById(id).orElse(null);
-    }
-
-    public Categorie findByNom(String nom) {
-        return categorieRepository.findByNom(nom);
     }
 
     public Categorie createOrUpdate(Categorie categorie) {
         return categorieRepository.save(categorie);
     }
 
-    public void delete(int id) {
+    public void delete(int id) throws ResourceNotFoundException {
+        if (!categorieRepository.existsById(id)) throw new ResourceNotFoundException("Category with id " + id + " not found.");
         categorieRepository.deleteById(id);
-    }
-
-    public void delete(Categorie categorie) {
-        categorieRepository.delete(categorie);
     }
 
 }
