+++ src/main/java/com/example/gestion-bibliotheque/service/LivreService.java
@@ -1,8 +1,7 @@
 package com.example.gestion_bibliotheque.service;
 
-import com.example.gestion_bibliotheque.entity.Auteur;
-import com.example.gestion_bibliotheque.entity.Categorie;
 import com.example.gestion_bibliotheque.entity.Livre;
+import com.example.gestion_bibliotheque.exception.ResourceNotFoundException;
 import com.example.gestion_bibliotheque.repository.LivreRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,32 +22,17 @@ public class LivreService {
         return livreRepository.findAll();
     }
 
-    public List<Livre> findAllByCategorie(Categorie categorie) {
-        return livreRepository.findAllByCategories(List.of(categorie));
-    }
-
-    public List<Livre> findAllByAuteur(Auteur auteur) {
-        return livreRepository.findAllByAuteur(auteur);
-    }
-
     public Livre findById(int id) {
         return livreRepository.findById(id).orElse(null);
-    }
-
-    public Livre findByTitre(String titre) {
-        return livreRepository.findByTitre(titre);
     }
 
     public Livre createOrUpdate(Livre livre) {
         return livreRepository.save(livre);
     }
 
-    public void delete(int id) {
+    public void delete(int id) throws ResourceNotFoundException {
+        if (!livreRepository.existsById(id)) throw new ResourceNotFoundException("Livre with id " + id + " not found.");
         livreRepository.deleteById(id);
-    }
-
-    public void delete(Livre livre) {
-        livreRepository.delete(livre);
     }
 
 }
