+++ src/main/java/com/example/gestion-bibliotheque/service/UtilisateurService.java
@@ -1,6 +1,7 @@
 package com.example.gestion_bibliotheque.service;
 
 import com.example.gestion_bibliotheque.entity.Utilisateur;
+import com.example.gestion_bibliotheque.exception.ResourceNotFoundException;
 import com.example.gestion_bibliotheque.repository.UtilisateurRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -25,24 +26,13 @@ public class UtilisateurService {
         return utilisateurRepository.findById(id).orElse(null);
     }
 
-    public Utilisateur findByNom(String nom) {
-        return utilisateurRepository.findByNom(nom);
-    }
-
-    public Utilisateur findByEmailAndMotDePasse(String email, String motDePasse) {
-        return utilisateurRepository.findByEmailAndMotDePasse(email, motDePasse);
-    }
-
     public Utilisateur createOrUpdate(Utilisateur utilisateur) {
         return utilisateurRepository.save(utilisateur);
     }
 
-    public void delete(Integer id) {
+    public void delete(Integer id) throws ResourceNotFoundException {
+        if (!utilisateurRepository.existsById(id)) throw new ResourceNotFoundException("User with id " + id + " not found.");
         utilisateurRepository.deleteById(id);
-    }
-
-    public void delete(Utilisateur utilisateur) {
-        utilisateurRepository.delete(utilisateur);
     }
 
 }
